# salesProcess/openapi.yaml
openapi: 3.0.0
info:
  title: Sales Process API
  description: Backend API for sales process management system with multi-step workflows
  version: 1.0.0
  contact:
    name: Project Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Advanced Dust Collection System
        description:
          type: string
          example: Advanced dust collection for cabinet manufacturing
        endUser:
          type: string
          example: Premier Cabinet Works
        country:
          type: string
          example: Germany
        industry:
          type: string
          enum: [woodworking, agroAndMilling, recycling, metalworking, paper, other]
          example: woodworking
        customIndustry:
          type: string
          nullable: true
        plantType:
          type: string
          example: Cabinet Manufacturing
        filterType:
          type: string
          example: Bag Filter
        fanType:
          type: string
          example: Centrifugal
        dustType:
          type: string
          example: Fine Wood Particles
        ductSystem:
          type: string
          example: Centralized Branch Network
        totalExtractionVolume:
          type: integer
          example: 15000
        pressure:
          type: integer
          example: 1200
        volumeFlow:
          type: integer
          example: 25000
        phoneNumber:
          type: string
          nullable: true
        privacySettings:
          type: object
          nullable: true
        processId:
          type: integer
          example: 1
        customerId:
          type: integer
          example: 1
        salesManagerId:
          type: integer
          example: 2
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Case:
      type: object
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          description: HTML content from rich text editor
          example: "<h2>Initial Consultation</h2><p>Meeting scheduled...</p>"
        processId:
          type: integer
          nullable: true
          example: 1
        referenceId:
          type: integer
          nullable: true
          example: 1
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /sales:
    post:
      summary: Create new sale
      tags:
        - Sales
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - endUser
                - country
                - industry
                - plantType
                - filterType
                - fanType
                - dustType
                - ductSystem
                - totalExtractionVolume
                - pressure
                - volumeFlow
                - customerId
              properties:
                title:
                  type: string
                description:
                  type: string
                endUser:
                  type: string
                country:
                  type: string
                industry:
                  type: string
                  enum: [woodworking, agroAndMilling, recycling, metalworking, paper, other]
                customIndustry:
                  type: string
                plantType:
                  type: string
                filterType:
                  type: string
                fanType:
                  type: string
                dustType:
                  type: string
                ductSystem:
                  type: string
                totalExtractionVolume:
                  type: integer
                pressure:
                  type: integer
                volumeFlow:
                  type: integer
                phoneNumber:
                  type: string
                customerId:
                  type: integer
      responses:
        '201':
          description: Sale created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sale'
        '403':
          description: Forbidden (requires salesManager role)
...
  /cases:
    post:
      summary: Create new case
      tags:
        - Cases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - processId
              properties:
                content:
                  type: string
                  description: HTML content from rich text editor
                processId:
                  type: integer
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Case'
        '403':
          description: Forbidden (requires Step 6 access)
...